UX-373 Rename systemProp helpers to props, Add clean helper function
