+++ packages/matchbox/src/components/Button/Button.js
@@ -6,7 +6,7 @@ import styled from 'styled-components';
 import { margin, width, padding, compose } from 'styled-system';
 import { createPropTypes } from '@styled-system/prop-types';
 import { omit } from '@styled-system/props';
-import { pick } from '../../helpers/systemProps';
+import { pick } from '../../helpers/props';
 import { Box } from '../Box';
 import { Spinner } from '../Spinner';
 import Icon from './Icon';
