+++ packages/matchbox/src/components/Modal/Content.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { padding } from 'styled-system';
 import { createPropTypes } from '@styled-system/prop-types';
-import { pick } from '../../helpers/systemProps';
+import { pick } from '../../helpers/props';
 
 import { Box } from '../Box';
