+++ packages/matchbox/src/components/RadioCard/Group.js
@@ -10,7 +10,7 @@ import { Label } from '../Label';
 import { OptionalLabel } from '../OptionalLabel';
 import { Stack } from '../Stack';
 
-import { pick } from '../../helpers/systemProps';
+import { pick } from '../../helpers/props';
 
 const Fieldset = styled.fieldset`
   border: none;
