+++ packages/matchbox/src/components/Select/Select.js
@@ -12,7 +12,7 @@ import { HelpText } from '../HelpText';
 import { compose, margin, maxWidth } from 'styled-system';
 import { createPropTypes } from '@styled-system/prop-types';
 import { omit } from '@styled-system/props';
-import { pick } from '../../helpers/systemProps';
+import { pick } from '../../helpers/props';
 import useInputDescribedBy from '../../hooks/useInputDescribedBy';
 import { focusOutline } from '../../styles/helpers';
