+++ packages/matchbox/src/components/Table/TableElements.js
@@ -5,7 +5,7 @@ import { cell, row, headerCell, totalsRow, verticalAlignment } from './styles';
 import { TablePaddingContext } from './context';
 import { padding, fontSize, compose } from 'styled-system';
 import { createPropTypes } from '@styled-system/prop-types';
-import { pick } from '../../helpers/systemProps';
+import { pick } from '../../helpers/props';
 
 const paddingAndFontSize = compose(padding, fontSize);
 const StyledCell = styled('td')`
